requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
cvzone
numpy
python-dotenv
requests
gunicorn
